Add files via upload
This is the final assessment i.e. a case study of the TechSaksham Training. The repository includes the dataset used in the case study and the code of the case study. In the case study, various factors were compared and graphs have been drawn out of them. The dataset chosen had various numeric factors of diabetes, BMI, pregnancies, glucose, etc. This case study and the graphs drawn out of them helps understanding the factors better and deduces the clear information out of those, handling the missing values and the outliers.